Rename chess/piece ==> chess/piecemoves tests... Resolves #15
