+++ shared/src/test/java/passoff/chess/piecemoves/PawnMoveTests.java
@@ -1,4 +1,4 @@
-package passoff.chess.piece;
+package passoff.chess.piecemoves;
 
 import chess.ChessMove;
 import chess.ChessPiece;
