+++ shared/src/test/java/passoff/chess/piecemoves/QueenMoveTests.java
@@ -1,4 +1,4 @@
-package passoff.chess.piece;
+package passoff.chess.piecemoves;
 
 import chess.ChessPosition;
 import org.junit.jupiter.api.Test;
